Finished writing all the other classes -- pending testing and debugging

+++ src/main/java/APriori.java
@@ -18,7 +18,7 @@ public class APriori {
     }
 
     //APriori algorithm,
-    public Set<Set<String>> getFrequentItemSets (int frequencyThreshold){
+    public Set<Set<String>> getFrequentItemSets (double frequencyThreshold){
         int k = 1;
         Set<Set<String>> union = new HashSet<>();
